Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int n;
-    int p;
-    int product=n*p;
+    float n;
+    float p;
+    float product=n*p;
     scanf("%f %f",&n,&p);
     printf("Product: %f",product);
     return 0;
